email-agent/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
anthropic>=0.40.0
rich>=13.0.0
textual>=0.50.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
promptfoo>=0.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
